Crate dafny_runtime

Re-exports

pub use itertools;

Modules

_System

array

dafny_runtime_conversions

object

rcmut

Macros

Extends

INIT_ARRAY_DATA

UpcastDef

UpcastDefObject

UpcastFn

UpcastObjectFn

array

cast

cast_any

cast_any_object

cast_object

int

is

is_object

map

maybe_placebos_from

md

modify

multiset

rd

read

refcount

seq

set

truncate

tuple_extract_index

update_field_if_uninit

update_field_if_uninit_object

update_field_nodrop

In that case, prefer to use update_uninit

update_field_nodrop_object

update_field_uninit

update_field_uninit_object

update_nodrop

update_nodrop_object

Structs

AllocationTracker

Array2

Array3

Array4

Array5

Array6

Array7

Array8

Array9

Array10

Array11

Array12

Array13

Array14

Array15

Array16

BigInt

A big signed integer type.

DafnyChar

DafnyCharUTF16

DafnyInt

DafnyPrintWrapper

ExactPool

FunctionWrapper

IntegerRange

IntegerRangeDown

IntegerRangeDownUnbounded

IntegerRangeUnbounded

Lazy

A value which is initialized on the first access.

LazyFieldWrapper

Map

MapBuilder

MaybePlacebo

Multiset

Object

Ptr

SequenceIter

Set

SetBuilder

Enums

Sequence

Traits

DafnyPrint

DafnyType

DafnyTypeEq

DafnyUsize

FromPrimitive

A generic trait for converting a number to a value.

Into

A value-to-value conversion that consumes the input value. The opposite of From.

NontrivialDefault

NumCast

An interface for casting between machine scalars.

ToPrimitive

A generic trait for converting a value to a number.

Upcast

UpcastObject

Zero

Defines an additive identity element for Self.

Functions

allocate

allocate_object

allocate_object_track

box_coerce

char_lt

dafny_rational_to_int

deallocate

euclidian_division

euclidian_modulo

exact_range

fn1_coerce

integer_range

integer_range_down

integer_range_down_unbounded

integer_range_unbounded

nullable_referential_equality

rc_coerce

string_of

string_utf16_of

transmute_unchecked⚠

upcast

upcast_id

upcast_object

Type Aliases

BigRational

Alias for arbitrary precision rationals.

DafnyString

DafnyStringUTF16

SizeT

Unions

MaybeUninit

A wrapper type to construct uninitialized instances of T.